Move stuff into the library
