+++ packages/input-composer/react/package.json
@@ -1,0 +1,4 @@
+{
+  "main": "dist/input-composer-react.cjs.js",
+  "module": "dist/input-composer-react.esm.js"
+}
