+++ packages/input-composer/src/react/index.ts
@@ -1,0 +1,2 @@
+export * from "./useGamepadInput"
+export * from "./useKeyboardInput"
